pyproject.toml
[tool.poetry]
name = "module5"
version = "0.1.0"
description = ""
authors = ["Дарья Рябова <doratraveled01@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
docutils = "^0.20.1"
drf-spectacular = "^0.26.3"
flake8 = "^6.0.0"
flake8-docstrings = "^1.7.0"
jsonschema = "^4.18.4"
pillow = "^10.0.0"
django-debug-toolbar = "^4.1.0"
sentry-sdk = "^1.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
